Restore the original behavior of a transparent background
With #9 merged the default behavior changed to rendering onto a white background, which now
is covered by an additional optional `options.background` parameter to the transform function.

While there:
* Fix the unit tests to pass again
* Improve typing and factor out the sharp usage into a helper for better readability

Note that the expected image changed slightly, but a visual comparison seems this is mostly
metadata and invisible artifacts from the svg2png -> sharp move.